requirements.txt
gunicorn
flask
flask-restful
Flask-SQLAlchemy
marshmallow-sqlalchemy
python-dotenv
psycopg2-binary
pytest
pydevd-pycharm~=221.5080.212
flask-marshmallow
Flask-Migrate
flask-jwt-extended
flask-admin